+++ athena/tradingtools/__init__.py
@@ -1,4 +1,5 @@
-from athena.tradingtools.models import Portfolio, Position
 from athena.tradingtools.strategies.strategy import Strategy
 
-__all__ = ["Strategy", "Position", "Portfolio"]
+__all__ = [
+    "Strategy",
+]
