+++ tests/core/test_market_entities.py
@@ -4,8 +4,8 @@ import pytest
 
 from athena.core.interfaces import Candle
 from athena.core.types import Coin
-from athena.tradingtools import Portfolio
-from athena.tradingtools.models import Position
+from athena.core.market_entities import Portfolio
+from athena.core.market_entities import Position
 
 
 @pytest.fixture
