+++ tests/performance/test_trading_session.py
@@ -4,7 +4,8 @@ import pytest
 
 from athena.core.interfaces import Fluctuations
 from athena.core.types import Signal
-from athena.tradingtools import Strategy, Portfolio
+from athena.tradingtools import Strategy
+from athena.core.market_entities import Portfolio
 from athena.performance.config import DataConfig
 from athena.performance.trading_session import TradingSession
